github: unify the security contact at the github supported policy document
The reason for moving the email addresses over is that the security policy
(`SECURITY.md`) is auto-link from a view places by github itself and should
therefore contain the primary contacts.

Further I removed the "supported versions" table which seems to be a
copy-pasta from a default until we define a clear guide here.